move mask_color_card to detect_color_card.py
utilize the same helper within the same function file rather than across the subpackage
